connection database and liquibase

+++ src/main/java/ru/skypro/homework/dto/Comment.java
@@ -7,7 +7,7 @@ public class Comment {
     private int author;
     private String authorImage;
     private String authorFirstName;
-    private Integer createdAt;
+    private long createdAt;
     private int pk;
     private String text;
 }
